Add files via upload
Fixed the errors (I think). As of right now, I am not to sure what works. Look through OI class to see which buttons do what. As for the PID controllers(subsystems and commands), not sure if they work the way I think so also we'll need to look at those.