Add files via upload
Se is the Smooth Autonomous (hence the S) program, but it is incomplete so the switch statement in robot.java has errors. I recommend copying C9's robot.java code into Se to make a testable S1.

PlotSmoothPath gives a runner class to look at paths on the field and speeds for the path. This is to be used when setting up autonomous paths so that construction is very quick with minimal testing needed.

What is needed for the new autonomous to work is knowing what the speed in ft/s of the robot is when at full power (1.0 speed controller).